# docs/api/openapi-unified.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MultiMind Unified Multi-Modal API",
    "description": "Multi-modal processing (text, image, audio) via Mixture-of-Experts or the multi-modal router. Provider API keys are read from the environment at request time; optional client authentication uses the `X-API-Key` header when the `API_KEYS` environment variable is set.",
    "version": "0.3.0"
  },
  "paths": {
    "/v1/process": {
      "post": {
        "tags": [
          "processing"
        ],
        "summary": "Process Request",
        "description": "Process multi-modal request using either MoE or router.",
        "operationId": "process_request_v1_process_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnifiedRequest",
                "examples": [
                  {
                    "inputs": [
                      {
                        "content": "Summarize the benefits of RAG.",
                        "modality": "text"
                      }
                    ],
                    "use_moe": true
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (e.g. no experts available)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/models": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "List Models",
        "description": "List available models and their capabilities.",
        "operationId": "list_models_v1_models_get",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Invalid request (e.g. no experts available)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workflows": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "List Workflows",
        "description": "List available MCP workflows.",
        "operationId": "list_workflows_v1_workflows_get",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Invalid request (e.g. no experts available)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/metrics": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Get Metrics",
        "description": "Get performance metrics for models.",
        "operationId": "get_metrics_v1_metrics_get",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Invalid request (e.g. no experts available)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Health Check",
        "description": "Health check endpoint (no auth, no provider keys required).",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/ready": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Readiness Check",
        "description": "Readiness probe; router and workflow registry are created lazily.",
        "operationId": "readiness_check_ready_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "version": {
            "type": "string",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "status",
          "version"
        ],
        "title": "HealthResponse"
      },
      "ModalityInput": {
        "properties": {
          "content": {
            "title": "Content"
          },
          "modality": {
            "type": "string",
            "title": "Modality"
          }
        },
        "type": "object",
        "required": [
          "content",
          "modality"
        ],
        "title": "ModalityInput",
        "description": "Input for a specific modality."
      },
      "UnifiedRequest": {
        "properties": {
          "inputs": {
            "items": {
              "$ref": "#/components/schemas/ModalityInput"
            },
            "type": "array",
            "title": "Inputs"
          },
          "use_moe": {
            "type": "boolean",
            "title": "Use Moe",
            "description": "Whether to use MoE processing",
            "default": true
          },
          "constraints": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Constraints",
            "description": "Processing constraints (cost, latency, etc.)"
          },
          "workflow": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow",
            "description": "Optional MCP workflow to use"
          }
        },
        "type": "object",
        "required": [
          "inputs"
        ],
        "title": "UnifiedRequest",
        "description": "Unified request structure for multi-modal processing."
      },
      "UnifiedResponse": {
        "properties": {
          "outputs": {
            "additionalProperties": true,
            "type": "object",
            "title": "Outputs"
          },
          "expert_weights": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expert Weights"
          },
          "metrics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metrics"
          }
        },
        "type": "object",
        "required": [
          "outputs"
        ],
        "title": "UnifiedResponse",
        "description": "Unified response structure."
      }
    }
  },
  "tags": [
    {
      "name": "processing",
      "description": "Multi-modal MoE / router processing"
    },
    {
      "name": "catalog",
      "description": "Model, workflow, and metrics discovery"
    },
    {
      "name": "system",
      "description": "Health and readiness probes"
    }
  ]
}